+++ UnitTests/CommonTestUtils/CommonTestUtils.csproj
@@ -35,9 +35,13 @@
   </PropertyGroup>
   <ItemGroup>
     <Reference Include="System" />
+    <Reference Include="System.IO.Compression.FileSystem" />
     <Reference Include="System.Windows.Forms" />
   </ItemGroup>
   <ItemGroup>
+    <PackageReference Include="LibGit2Sharp">
+      <Version>0.25.0</Version>
+    </PackageReference>
     <PackageReference Include="NUnit" Version="3.10.1" />
   </ItemGroup>
   <ItemGroup>
@@ -45,6 +49,10 @@
     <Compile Include="EmbeddedResourceLoader.cs" />
     <Compile Include="GitModuleTestHelper.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="TestRepository\WellKnown\CodedTestRepository.cs" />
+    <Compile Include="TestRepository\ITestRepositoryData.cs" />
+    <Compile Include="TestRepository\TestRepository.cs" />
+    <Compile Include="TestRepository\ZippedTestRepoDirectory.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\GitCommands\GitCommands.csproj">
