+++ UnitTests/CommonTestUtils/TestRepository/ITestRepositoryData.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace CommonTestUtils.TestRepository
+{
+    public interface ITestRepositoryData : IDisposable
+    {
+        string ContentPath { get; }
+    }
+}
