+++ UnitTests/GitCommandIntegrationTests/GitCommandIntegrationTests.csproj
@@ -46,13 +46,9 @@
     <Reference Include="System.IO.Compression.FileSystem" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="CodedTestRepository.cs" />
     <Compile Include="CodedRepositoryPoC.cs" />
-    <Compile Include="TestRepository.cs" />
-    <Compile Include="ITestRepositoryData.cs" />
     <Compile Include="PoC.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="ZippedTestRepoDirectory.cs" />
   </ItemGroup>
   <ItemGroup>
     <PackageReference Include="ApprovalTests">
@@ -60,9 +56,6 @@
     </PackageReference>
     <PackageReference Include="FluentAssertions">
       <Version>5.2.0</Version>
-    </PackageReference>
-    <PackageReference Include="LibGit2Sharp">
-      <Version>0.25.0</Version>
     </PackageReference>
     <PackageReference Include="NSubstitute">
       <Version>3.1.0</Version>
