--- UnitTests/GitCommandIntegrationTests/ITestRepositoryData.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace GitCommandIntegrationTests
-{
-    internal interface ITestRepositoryData : IDisposable
-    {
-        string ContentPath { get; }
-    }
-}
