+++ UnitTests/GitCommandIntegrationTests/PoC.cs
@@ -1,4 +1,5 @@
 ﻿using ApprovalTests;
+using CommonTestUtils.TestRepository;
 using GitCommands;
 using NUnit.Framework;
